feat:create and consume AppContext
